Refactor into subcommands:
