+++ packages/flutter_tools/lib/src/commands/migrate_abandon.dart
@@ -1,0 +1,62 @@
+// Copyright 2014 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:meta/meta.dart';
+
+import '../base/file_system.dart';
+import '../base/logger.dart';
+import '../globals.dart' as globals;
+import '../project.dart';
+import '../runner/flutter_command.dart';
+import '../migrate/migrate_config.dart';
+import '../migrate/migrate_manifest.dart';
+import '../migrate/migrate_utils.dart';
+import '../cache.dart';
+import 'migrate.dart';
+
+class MigrateAbandonCommand extends FlutterCommand {
+  MigrateAbandonCommand({
+    bool verbose = false,
+  }) : _verbose = verbose {
+    requiresPubspecYaml();
+    argParser.addOption(
+      'working-directory',
+      help: '',
+      defaultsTo: null,
+      valueHelp: 'path',
+    );
+  }
+
+  final bool _verbose;
+
+  @override
+  final String name = 'abandon';
+
+  @override
+  final String description = 'Deletes the current migration working directory';
+
+  @override
+  String get category => FlutterCommandCategory.project;
+
+  @override
+  Future<Set<DevelopmentArtifact>> get requiredArtifacts async => const <DevelopmentArtifact>{};
+
+  @override
+  Future<FlutterCommandResult> runCommand() async {
+    Directory workingDir = FlutterProject.current().directory.childDirectory(kDefaultMigrateWorkingDirectoryName);
+    if (stringArg('working-directory') != null) {
+      workingDir = globals.fs.directory(stringArg('working-directory'));
+    }
+    if (!workingDir.existsSync()) {
+      print('No migration in progress. Start a new migration with:\n');
+      print('    \$ flutter migrate start\n');
+      return const FlutterCommandResult(ExitStatus.fail);
+    }
+    print('Abandoning the existing migration will delete the migration working directory at ${workingDir.path}');
+    workingDir.deleteSync(recursive: true);
+    print('Abandon complete. Start a new migration with:\n');
+    print('    \$ flutter migrate start\n');
+    return const FlutterCommandResult(ExitStatus.success);
+  }
+}
